requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
numpy==1.21.0
opencv-python==4.5.3.56
pyzbar==0.1.8